Create server extension to serve the custom order files
This way, the files can be hidden.  Jupyter won't serve the hidden
files by default, and I think enabling this will make them visible
in the directory view.  We don't want that, so we'll set up the
server extension.  This also will give us flexibility in the future
for how to create the order.